requirements.txt
flask
requests
beautifulsoup4
google-generativeai
mysql-connector-python
gunicorn
